requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

sqlalchemy[asyncio]==2.0.35
asyncmy==0.2.9
alembic==1.13.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

httpx==0.27.2
websockets==12.0
python-socketio==5.11.4

structlog==24.4.0
prometheus-client==0.21.0
prometheus-fastapi-instrumentator==7.0.0

opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
pytest==8.3.3
anyio==4.4.0
aiosqlite==0.21.0
